requirements.txt
streamlit
chromadb
sentence-transformers
torch
transformers
python-dotenv
tqdm
numpy
pandas
tiktoken
PyPDF2